requirements.txt
langchain
langchain_community
youtube_transcript_api
openai
gradio
elevenlabs
pyyaml